# coffee-be/static/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ApiResponse_any_": {
				"properties": {
					"data": {
						"additionalProperties": true
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChainBulkPayment": {
				"description": "Chain bulk payment (when paying at the bank to multiple farmers)",
				"properties": {
					"docType": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string",
						"nullable": true
					},
					"_rev": {
						"type": "string",
						"nullable": true
					},
					"created": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of creation"
					},
					"lastChange": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of last change"
					},
					"userCreatedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that created the document."
					},
					"userChangedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that changed the document."
					},
					"currency": {
						"type": "string",
						"description": "Currency of payment"
					},
					"formalCreationTime": {
						"type": "string",
						"description": "Formal creation time set while entering."
					},
					"bankInfo": {
						"$ref": "#/components/schemas/BankAccountInfo",
						"description": "Bank account info where transfer iz carried out."
					},
					"payingOrganizationId": {
						"type": "string",
						"description": "Organization that is paying"
					},
					"paymentPurposeType": {
						"anyOf": [
							{
								"type": "string",
								"enum": [
									"ADVANCE_PAYMENT"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"FIRST_INSTALLMENT"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"SECOND_INSTALLMENT"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"WOMEN_PREMIUM"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"INVOICE_PAYMENT"
								],
								"nullable": false
							}
						],
						"description": "Payment purpose type."
					},
					"paymentDescription": {
						"type": "string",
						"description": "Payment description."
					},
					"totalAmount": {
						"type": "number",
						"format": "double",
						"description": "Total amount of payment in currency"
					},
					"paymentPerKg": {
						"type": "number",
						"format": "double",
						"description": "Payment per kg in currency"
					},
					"additionalCost": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Additional cost of payment in currency"
					},
					"additionalCostDescription": {
						"type": "string",
						"nullable": true,
						"description": "Additional cost description."
					},
					"stockOrderIds": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "List of stockOrder ids."
					},
					"additionalProofs": {
						"items": {
							"$ref": "#/components/schemas/ChainActivityProof"
						},
						"type": "array",
						"description": "List of additional proofs."
					},
					"receiptNumber": {
						"type": "string",
						"nullable": true,
						"description": "Receipt number"
					},
					"payingOrganization": {
						"$ref": "#/components/schemas/ChainOrganization",
						"description": "Organization that is paying. Not stored. On write _id is used."
					},
					"payments": {
						"items": {
							"$ref": "#/components/schemas/ChainPayment"
						},
						"type": "array",
						"description": "List of payments."
					}
				},
				"required": [
					"currency",
					"formalCreationTime",
					"payingOrganizationId",
					"paymentDescription",
					"totalAmount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChainPayment": {
				"description": "Chain payment",
				"properties": {
					"docType": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string",
						"nullable": true
					},
					"_rev": {
						"type": "string",
						"nullable": true
					},
					"created": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of creation"
					},
					"lastChange": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of last change"
					},
					"userCreatedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that created the document."
					},
					"userChangedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that changed the document."
					},
					"formalCreationTime": {
						"type": "string",
						"description": "Formal creation time set while entering."
					},
					"paymentType": {
						"anyOf": [
							{
								"type": "string",
								"enum": [
									"CASH"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"BANK"
								],
								"nullable": false
							}
						],
						"description": "Type of payment."
					},
					"currency": {
						"type": "string",
						"description": "Currency of payment"
					},
					"amount": {
						"type": "number",
						"format": "double",
						"description": "Amount of payment in currency"
					},
					"amountPaidToTheCollector": {
						"type": "number",
						"format": "double",
						"description": "Amount paid to the collector"
					},
					"stockOrderId": {
						"type": "string",
						"description": "Stock order for which payment was done (order reference)."
					},
					"orderId": {
						"type": "string",
						"description": "Order for which payment was done (when payment purpose is second_installment aka member bonus)."
					},
					"transactionIds": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Reference to (input) transactions for which payment was actually done."
					},
					"payingOrganizationId": {
						"type": "string",
						"description": "Organization that is paying"
					},
					"recipientOrganizationId": {
						"type": "string",
						"nullable": true,
						"description": "Recipient organization"
					},
					"recipientUserCustomerId": {
						"type": "string",
						"nullable": true,
						"description": "Recipient user customer"
					},
					"representativeOfRecipientOrganizationId": {
						"type": "string",
						"nullable": true,
						"description": "Recipient organization (i.e. collector's organization)"
					},
					"representativeOfRecipientUserCustomerId": {
						"type": "string",
						"nullable": true,
						"description": "Recipient user customer (i.e. collector)"
					},
					"recipientCompanyCustomerId": {
						"type": "string",
						"nullable": true,
						"description": "Recipient company  customer"
					},
					"recipientType": {
						"anyOf": [
							{
								"type": "string",
								"enum": [
									"ORGANIZATION"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"COMPANY_CUSTOMER"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"USER_CUSTOMER"
								],
								"nullable": false
							}
						],
						"description": "Type of recipient."
					},
					"receiptNumber": {
						"type": "string",
						"nullable": true,
						"description": "Receipt number on the attached receipt (receiptDocumentId). Mandatory for the cash transfer."
					},
					"receiptDocument": {
						"$ref": "#/components/schemas/ChainFileInfo",
						"description": "Receipt document. Required for the cash transfer."
					},
					"receiptDocumentType": {
						"anyOf": [
							{
								"type": "string",
								"enum": [
									"PURCHASE_SHEET"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"RECEIPT"
								],
								"nullable": false
							}
						],
						"description": "Receipt document type."
					},
					"bankTransferId": {
						"type": "string",
						"nullable": true,
						"description": "Reference to the bulk payment."
					},
					"paymentPurposeType": {
						"anyOf": [
							{
								"type": "string",
								"enum": [
									"ADVANCE_PAYMENT"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"FIRST_INSTALLMENT"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"SECOND_INSTALLMENT"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"WOMEN_PREMIUM"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"INVOICE_PAYMENT"
								],
								"nullable": false
							}
						],
						"description": "Payment purpose type."
					},
					"paymentStatus": {
						"anyOf": [
							{
								"type": "string",
								"enum": [
									"UNCONFIRMED"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"CONFIRMED"
								],
								"nullable": false
							}
						],
						"description": "Payment status."
					},
					"paymentConfirmedByUser": {
						"type": "string",
						"nullable": true,
						"description": "Payment confirmed by userId."
					},
					"paymentConfirmedByOrganization": {
						"type": "string",
						"nullable": true,
						"description": "Payment confirmed by oganizationId."
					},
					"paymentConfirmedAtTime": {
						"type": "string",
						"nullable": true,
						"description": "Payment confirmed at this time."
					},
					"payingOrganization": {
						"$ref": "#/components/schemas/ChainOrganization",
						"description": "Organization that is paying. Not stored. On write _id is used."
					},
					"recipientOrganization": {
						"$ref": "#/components/schemas/ChainOrganization",
						"description": "Recipient organization. Not stored. On write _id is used."
					},
					"recipientUserCustomer": {
						"$ref": "#/components/schemas/ChainUserCustomer",
						"description": "Recipient user customer. Not stored. On write _id is used."
					},
					"recipientCompanyCustomer": {
						"$ref": "#/components/schemas/ChainUserCustomer",
						"description": "Recipient user customer. Not stored. On write _id is used."
					},
					"bankTransfer": {
						"$ref": "#/components/schemas/ChainBulkPayment",
						"description": "Bank transfer. Not stored. On write _id is used."
					},
					"representativeOfRecipientOrganization": {
						"$ref": "#/components/schemas/ChainOrganization",
						"description": "Recipient organization (i.e. collector's organization). Not stored. On write _id is used."
					},
					"representativeOfRecipientUserCustomer": {
						"$ref": "#/components/schemas/ChainUserCustomer",
						"description": "Recipient user customer (i.e. collector). Not stored. On write _id is used."
					},
					"queryFacilityName": {
						"type": "string",
						"nullable": true,
						"description": "Stock order facility"
					},
					"queryPurchaseOrderName": {
						"type": "string",
						"nullable": true,
						"description": "Stock order name."
					},
					"queryProducerUserCustomerName": {
						"type": "string",
						"nullable": true,
						"description": "Producer user customer name."
					},
					"preferredWayOfPayment": {
						"type": "string",
						"nullable": true
					},
					"productionDate": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"formalCreationTime",
					"paymentType",
					"currency",
					"amount",
					"stockOrderId",
					"orderId",
					"payingOrganizationId",
					"recipientType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ChainBulkPayment_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ChainBulkPayment"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_PaginatedList_ChainBulkPayment__": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PaginatedList_ChainBulkPayment_"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ChainPayment_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ChainPayment"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_PaginatedList_ChainPayment__": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PaginatedList_ChainPayment_"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "coffee-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/chain-api/data/payment/bulk-payment/{dbId}": {
			"get": {
				"operationId": "GetBulkPayment",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainBulkPayment_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches a bulk payment by its _id.",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"description": "database id of the bulk payment",
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/data/payment/bulk-payment": {
			"post": {
				"operationId": "PostBulkPayment",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates a bulk payment. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainBulkPayment"
							}
						}
					}
				}
			}
		},
		"/chain-api/data/payment/bulk-payment/list/organization/{payingOrganizationId}": {
			"get": {
				"operationId": "ListBulkPaymentsForPayingOrganization",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainBulkPayment__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Returns paginated list of bulk payments for payingOrganization",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "payingOrganizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/payment/delete": {
			"post": {
				"operationId": "DeletePayment",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes payment",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainPayment"
							}
						}
					}
				}
			}
		},
		"/chain-api/data/payment/{dbId}": {
			"get": {
				"operationId": "GetPayment",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainPayment_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches a payment by its _id.",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"description": "database id of the payment",
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/data/payment": {
			"post": {
				"operationId": "PostPayment",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates a payment. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainPayment"
							}
						}
					}
				}
			}
		},
		"/chain-api/data/payment/list/organization/{payingOrganizationId}": {
			"get": {
				"operationId": "ListPaymentsForPayingOrganization",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainPayment__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of payments",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "payingOrganizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "query by receiptNumber",
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "PAYMENT_DATE or DELIVERY_DATE",
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {}
					},
					{
						"description": "CONFIRMED or UNCONFIRMED",
						"in": "query",
						"name": "paymentStatus",
						"required": false,
						"schema": {}
					},
					{
						"description": "CASH_VIA_COOPERATIVE or CASH_VIA_COLLECTOR or BANK_TRANSFER or UNKNOWN",
						"in": "query",
						"name": "wayOfPayment",
						"required": false,
						"schema": {}
					},
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "deliveryDateStart",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "deliveryDateEnd",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/data/payment/list/stock-order/{stockOrderId}": {
			"get": {
				"operationId": "ListPaymentsForStockOrder",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainPayment__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of payments for a given stock order",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"description": "_id of a given organization",
						"in": "path",
						"name": "stockOrderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/payment/list/farmer/{farmerId}": {
			"get": {
				"operationId": "ListPaymentsForRecipientUserCustomer",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainPayment__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of payments for a user customer",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "farmerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/payment/list/bank-transfer/{bankTransferId}": {
			"get": {
				"operationId": "ListPaymentsForBankTransfer",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainPayment__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of payments for a given",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bankTransferId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/payment/confirm-payment": {
			"post": {
				"operationId": "ConfirmPayment",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Confirm payment",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainPayment"
							}
						}
					}
				}
			}
		},
...
		"/chain-api/data/user-customer/list/payments/{userCustomerId}": {
			"get": {
				"operationId": "ListPaymentsorUserCustomer",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainPayment__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of payments for a given user customer",
				"tags": [
					"User Customer"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userCustomerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}